BREAKING: Remove library module, move types/config to src/
Move src/library/{types,config,config_loader}.rs directly under src/.
Delete src/library/mod.rs. Update imports in lib.rs and tests.
Update copilot-instructions.md and roadmap.md file references.
