--- src/library/mod.rs
@@ -1,8 +1,0 @@
-//! Library module — core types, configuration, and config loading.
-//!
-//! This module defines the foundational type system for the factory manager
-//! game: tokens, cards, effects, contracts, and game rules.
-
-pub mod config;
-pub mod config_loader;
-pub mod types;
